meta, scripts I started, prefabs, and animations
none of which I got to completion. Conveyor belt should be done once I
finished
more on this update:

Started working on a script that tells the degree of the dogs
conditioning, as well as animations to show each state. Got the
conveyor belt to sound a bell (I think no sound to test yet) and then
make an explosion to scare the dog. Now all we have to do is show how
this is conditioning the dogs on the conveyor belt